requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
neomodel>=4.0.8
python-dotenv>=0.21.0
pytest>=7.2.2
httpx>=0.24.0
